test: extract skill injector suites
