+++ crates/agent-engine/src/skills/injector.rs
@@ -1205,1286 +1205,13 @@ pub fn render_skill_unavailable_with_remediation(
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_extract_mentions() {
-        assert_eq!(
-            extract_mentions("Use $test-skill for testing"),
-            vec!["test-skill"]
-        );
-
-        assert!(extract_mentions("Run $my_skill please").is_empty());
-
-        // Multiple mentions
-        let mentions = extract_mentions("$skill-a and $skill-b");
-        assert_eq!(mentions, vec!["skill-a", "skill-b"]);
-
-        // With punctuation at end
-        assert_eq!(extract_mentions("Use $skill-name."), vec!["skill-name"]);
-
-        // No mentions
-        assert!(extract_mentions("Plain text without mentions").is_empty());
-    }
-
-    #[test]
-    fn test_inject_skills() {
-        let skill = Skill {
-            metadata: SkillMetadata {
-                id: "test".to_string(),
-                package_id: None,
-                name: "Test Skill".to_string(),
-                description: "A test".to_string(),
-                short_description: None,
-                path: std::path::PathBuf::from("/tmp/test/SKILL.md"),
-                package_root: None,
-                resource_root: None,
-                scope: SkillScope::Installed,
-                tags: vec![],
-                capabilities: None,
-                compatibility: Default::default(),
-                source: SkillContentSource::File(std::path::PathBuf::from("/tmp/test/SKILL.md")),
-                enabled: true,
-                allow_implicit_invocation: true,
-                alan_metadata: Default::default(),
-                compatible_metadata: Default::default(),
-                execution: Default::default(),
-            },
-            content: "# Instructions\n\nDo this and that.".to_string(),
-            frontmatter: SkillFrontmatter {
-                name: "Test Skill".to_string(),
-                description: "A test".to_string(),
-                metadata: Default::default(),
-                capabilities: Default::default(),
-                compatibility: Default::default(),
-            },
-        };
-
-        let injected = inject_skills(&[skill]);
-        assert!(injected.contains("## Skill: Test Skill"));
-        assert!(injected.contains("# Instructions"));
-    }
-
-    #[test]
-    fn test_inject_skills_conservatively_falls_back_inline_without_runtime_support() {
-        let skill = Skill {
-            metadata: SkillMetadata {
-                id: "repo-review".to_string(),
-                package_id: Some("pkg:repo-review".to_string()),
-                name: "Repo Review".to_string(),
-                description: "Review repository changes".to_string(),
-                short_description: Some("Delegated review capability".to_string()),
-                path: std::path::PathBuf::from("/tmp/repo-review/SKILL.md"),
-                package_root: Some(std::path::PathBuf::from("/tmp/repo-review")),
-                resource_root: Some(std::path::PathBuf::from("/tmp/repo-review")),
-                scope: SkillScope::Installed,
-                tags: vec![],
-                capabilities: None,
-                compatibility: Default::default(),
-                source: SkillContentSource::File(std::path::PathBuf::from(
-                    "/tmp/repo-review/SKILL.md",
-                )),
-                enabled: true,
-                allow_implicit_invocation: true,
-                alan_metadata: Default::default(),
-                compatible_metadata: Default::default(),
-                execution: ResolvedSkillExecution::Delegate {
-                    target: "reviewer".to_string(),
-                    source: SkillExecutionResolutionSource::ExplicitMetadata,
-                },
-            },
-            content: "SECRET INLINE BODY".to_string(),
-            frontmatter: SkillFrontmatter {
-                name: "Repo Review".to_string(),
-                description: "Review repository changes".to_string(),
-                metadata: Default::default(),
-                capabilities: Default::default(),
-                compatibility: Default::default(),
-            },
-        };
-
-        let injected = inject_skills(&[skill]);
-        assert!(injected.contains("### Runtime Fallback"));
-        assert!(injected.contains("SECRET INLINE BODY"));
-        assert!(!injected.contains("### Delegated Capability"));
-    }
-
-    #[test]
-    fn test_render_active_skill_prompt_can_render_delegated_stub_with_runtime_support() {
-        let skill = Skill {
-            metadata: SkillMetadata {
-                id: "repo-review".to_string(),
-                package_id: Some("pkg:repo-review".to_string()),
-                name: "Repo Review".to_string(),
-                description: "Review repository changes".to_string(),
-                short_description: Some("Delegated review capability".to_string()),
-                path: std::path::PathBuf::from("/tmp/repo-review/SKILL.md"),
-                package_root: Some(std::path::PathBuf::from("/tmp/repo-review")),
-                resource_root: Some(std::path::PathBuf::from("/tmp/repo-review")),
-                scope: SkillScope::Installed,
-                tags: vec![],
-                capabilities: None,
-                compatibility: Default::default(),
-                source: SkillContentSource::File(std::path::PathBuf::from(
-                    "/tmp/repo-review/SKILL.md",
-                )),
-                enabled: true,
-                allow_implicit_invocation: true,
-                alan_metadata: Default::default(),
-                compatible_metadata: Default::default(),
-                execution: ResolvedSkillExecution::Delegate {
-                    target: "reviewer".to_string(),
-                    source: SkillExecutionResolutionSource::ExplicitMetadata,
-                },
-            },
-            content: "SECRET INLINE BODY".to_string(),
-            frontmatter: SkillFrontmatter {
-                name: "Repo Review".to_string(),
-                description: "Review repository changes".to_string(),
-                metadata: Default::default(),
-                capabilities: Default::default(),
-                compatibility: Default::default(),
-            },
-        };
-
-        let envelope = ActiveSkillEnvelope::available(
-            skill.metadata.clone(),
-            SkillActivationReason::ExplicitMention {
-                mention: "repo-review".to_string(),
-            },
-        );
-        let rendered = render_active_skill_prompt_for_runtime(&skill, &envelope, true).rendered;
-
-        assert!(rendered.contains("### Delegated Capability"));
-        assert!(rendered.contains("invoke_delegated_skill"));
-        assert!(rendered.contains("output_ref"));
-        assert!(rendered.contains("namespace file at `output_ref.path`"));
-        assert!(!rendered.contains("referenced child rollout/machine"));
-        assert!(rendered.contains("child_run"));
-        assert!(rendered.contains("\"target\": \"reviewer\""));
-        assert!(!rendered.contains("SECRET INLINE BODY"));
-    }
-
-    #[test]
-    fn test_inject_skills_renders_unresolved_stub_without_full_body() {
-        let skill = Skill {
-            metadata: SkillMetadata {
-                id: "skill-creator".to_string(),
-                package_id: Some("pkg:skill-creator".to_string()),
-                name: "Skill Creator".to_string(),
-                description: "Create and grade skills".to_string(),
-                short_description: None,
-                path: std::path::PathBuf::from("/tmp/skill-creator/SKILL.md"),
-                package_root: Some(std::path::PathBuf::from("/tmp/skill-creator")),
-                resource_root: Some(std::path::PathBuf::from("/tmp/skill-creator")),
-                scope: SkillScope::Installed,
-                tags: vec![],
-                capabilities: None,
-                compatibility: Default::default(),
-                source: SkillContentSource::File(std::path::PathBuf::from(
-                    "/tmp/skill-creator/SKILL.md",
-                )),
-                enabled: true,
-                allow_implicit_invocation: true,
-                alan_metadata: Default::default(),
-                compatible_metadata: Default::default(),
-                execution: ResolvedSkillExecution::Unresolved {
-                    reason: SkillExecutionUnresolvedReason::AmbiguousPackageShape {
-                        skill_id: "skill-creator".to_string(),
-                        child_agent_exports: vec![
-                            "creator".to_string(),
-                            "grader".to_string(),
-                            "analyzer".to_string(),
-                        ],
-                    },
-                },
-            },
-            content: "INLINE BODY SHOULD NOT APPEAR".to_string(),
-            frontmatter: SkillFrontmatter {
-                name: "Skill Creator".to_string(),
-                description: "Create and grade skills".to_string(),
-                metadata: Default::default(),
-                capabilities: Default::default(),
-                compatibility: Default::default(),
-            },
-        };
-
-        let injected = inject_skills(&[skill]);
-        assert!(injected.contains("### Skill Execution Status"));
-        assert!(injected.contains("reason: ambiguous_package_shape"));
-        assert!(injected.contains("child_agent_exports: creator, grader, analyzer"));
-        assert!(!injected.contains("INLINE BODY SHOULD NOT APPEAR"));
-    }
-
-    #[test]
-    fn test_build_prompt_with_skills() {
-        let skill = Skill {
-            metadata: SkillMetadata {
-                id: "eval".to_string(),
-                package_id: None,
-                name: "Evaluation".to_string(),
-                description: "Eval".to_string(),
-                short_description: None,
-                path: std::path::PathBuf::from("/tmp/eval/SKILL.md"),
-                package_root: None,
-                resource_root: None,
-                scope: SkillScope::Installed,
-                tags: vec![],
-                capabilities: None,
-                compatibility: Default::default(),
-                source: SkillContentSource::File(std::path::PathBuf::from("/tmp/eval/SKILL.md")),
-                enabled: true,
-                allow_implicit_invocation: true,
-                alan_metadata: Default::default(),
-                compatible_metadata: Default::default(),
-                execution: Default::default(),
-            },
-            content: "Follow these steps.".to_string(),
-            frontmatter: SkillFrontmatter {
-                name: "Evaluation".to_string(),
-                description: "Eval".to_string(),
-                metadata: Default::default(),
-                capabilities: Default::default(),
-                compatibility: Default::default(),
-            },
-        };
-
-        let prompt = build_prompt_with_skills("Evaluate this", &[skill]);
-        assert!(prompt.contains("Follow these steps"));
-        assert!(prompt.contains("Evaluate this"));
-    }
-
-    #[test]
-    fn test_render_skills_list() {
-        let skills = vec![
-            SkillMetadata {
-                id: "skill-a".to_string(),
-                package_id: None,
-                name: "Skill A".to_string(),
-                description: "Does A".to_string(),
-                short_description: Some("Short A".to_string()),
-                path: std::path::PathBuf::from("/a/SKILL.md"),
-                package_root: None,
-                resource_root: None,
-                scope: SkillScope::Installed,
-                tags: vec![],
-                capabilities: None,
-                compatibility: Default::default(),
-                source: SkillContentSource::File(std::path::PathBuf::from("/a/SKILL.md")),
-                enabled: true,
-                allow_implicit_invocation: true,
-                alan_metadata: Default::default(),
-                compatible_metadata: CompatibleSkillMetadata {
-                    interface: CompatibleSkillInterface {
-                        display_name: Some("UI Skill A".to_string()),
-                        short_description: Some("UI Short A".to_string()),
-                        ..Default::default()
-                    },
-                    ..Default::default()
-                },
-                execution: Default::default(),
-            },
-            SkillMetadata {
-                id: "skill-b".to_string(),
-                package_id: None,
-                name: "Skill B".to_string(),
-                description: "Does B".to_string(),
-                short_description: None,
-                path: std::path::PathBuf::from("/b/SKILL.md"),
-                package_root: None,
-                resource_root: None,
-                scope: SkillScope::Descriptor,
-                tags: vec![],
-                capabilities: None,
-                compatibility: Default::default(),
-                source: SkillContentSource::File(std::path::PathBuf::from("/b/SKILL.md")),
-                enabled: true,
-                allow_implicit_invocation: true,
-                alan_metadata: Default::default(),
-                compatible_metadata: Default::default(),
-                execution: Default::default(),
-            },
-            SkillMetadata {
-                id: "skill-c".to_string(),
-                package_id: None,
-                name: "Skill C".to_string(),
-                description: "Does C".to_string(),
-                short_description: None,
-                path: std::path::PathBuf::from("/c/SKILL.md"),
-                package_root: None,
-                resource_root: None,
-                scope: SkillScope::Descriptor,
-                tags: vec![],
-                capabilities: None,
-                compatibility: Default::default(),
-                source: SkillContentSource::File(std::path::PathBuf::from("/c/SKILL.md")),
-                enabled: true,
-                allow_implicit_invocation: true,
-                alan_metadata: Default::default(),
-                compatible_metadata: Default::default(),
-                execution: ResolvedSkillExecution::Delegate {
-                    target: "reviewer".to_string(),
-                    source: SkillExecutionResolutionSource::ExplicitMetadata,
-                },
-            },
-        ];
-
-        let list = render_skills_list(&skills, true).unwrap();
-        assert!(list.contains("## Available Skills"));
-        assert!(list.contains("- skill_id: skill-a"));
-        assert!(list.contains("  name: Skill A"));
-        assert!(list.contains("  description: Does A"));
-        assert!(!list.contains("UI Skill A"));
-        assert!(!list.contains("UI Short A"));
-        assert!(list.contains("  skill_path: /a/SKILL.md"));
-        assert!(list.contains("- skill_id: skill-b"));
-        assert!(list.contains("  description: Does B"));
-        assert!(list.contains("- skill_id: skill-c"));
-        assert!(list.contains("  execution: delegate(target=reviewer)"));
-        assert!(list.contains("  use: call `invoke_delegated_skill` directly"));
-        assert!(list.contains("Available Skills"));
-    }
-
-    #[test]
-    fn test_render_skills_list_falls_back_to_inline_guidance_without_delegated_support() {
-        let skills = vec![SkillMetadata {
-            id: "skill-c".to_string(),
-            package_id: None,
-            name: "Skill C".to_string(),
-            description: "Does C".to_string(),
-            short_description: None,
-            path: std::path::PathBuf::from("/c/SKILL.md"),
-            package_root: None,
-            resource_root: None,
-            scope: SkillScope::Descriptor,
-            tags: vec![],
-            capabilities: None,
-            compatibility: Default::default(),
-            source: SkillContentSource::File(std::path::PathBuf::from("/c/SKILL.md")),
-            enabled: true,
-            allow_implicit_invocation: true,
-            alan_metadata: Default::default(),
-            compatible_metadata: Default::default(),
-            execution: ResolvedSkillExecution::Delegate {
-                target: "reviewer".to_string(),
-                source: SkillExecutionResolutionSource::ExplicitMetadata,
-            },
-        }];
-
-        let list = render_skills_list(&skills, false).unwrap();
-        assert!(list.contains("  skill_path: /c/SKILL.md"));
-        assert!(!list.contains("invoke_delegated_skill"));
-        assert!(!list.contains("execution: delegate("));
-    }
-
-    #[test]
-    fn test_render_skills_list_hides_builtin_package_paths() {
-        let skills = vec![SkillMetadata {
-            id: "memory".to_string(),
-            package_id: Some("builtin:alan-memory".to_string()),
-            name: "Memory".to_string(),
-            description: "Persistent memory across Agent Processes".to_string(),
-            short_description: None,
-            path: std::path::PathBuf::from(
-                "/private/tmp/alan/builtin-skill-packages/0.1.0/123/memory/SKILL.md",
-            ),
-            package_root: Some(std::path::PathBuf::from(
-                "/private/tmp/alan/builtin-skill-packages/0.1.0/123/memory",
-            )),
-            resource_root: Some(std::path::PathBuf::from(
-                "/private/tmp/alan/builtin-skill-packages/0.1.0/123/memory",
-            )),
-            scope: SkillScope::Builtin,
-            tags: vec![],
-            capabilities: None,
-            compatibility: Default::default(),
-            source: SkillContentSource::File(std::path::PathBuf::from(
-                "/private/tmp/alan/builtin-skill-packages/0.1.0/123/memory/SKILL.md",
-            )),
-            enabled: true,
-            allow_implicit_invocation: true,
-            alan_metadata: Default::default(),
-            compatible_metadata: Default::default(),
-            execution: Default::default(),
-        }];
-
-        let list = render_skills_list(&skills, false).unwrap();
-        assert!(list.contains("  skill_source: builtin_capability_package"));
-        assert!(list.contains("rely on the runtime-disclosed instructions"));
-        assert!(!list.contains("skill_path:"));
-        assert!(!list.contains("builtin-skill-packages"));
-    }
-
-    #[test]
-    fn test_render_skills_list_keeps_builtin_delegated_target_guidance() {
-        let skills = vec![SkillMetadata {
-            id: "skill-creator".to_string(),
-            package_id: Some("builtin:alan-skill-creator".to_string()),
-            name: "Skill Creator".to_string(),
-            description: "Create or update alan skill packages".to_string(),
-            short_description: None,
-            path: std::path::PathBuf::from(
-                "/private/tmp/alan/builtin-skill-packages/0.1.0/123/skill-creator/SKILL.md",
-            ),
-            package_root: Some(std::path::PathBuf::from(
-                "/private/tmp/alan/builtin-skill-packages/0.1.0/123/skill-creator",
-            )),
-            resource_root: Some(std::path::PathBuf::from(
-                "/private/tmp/alan/builtin-skill-packages/0.1.0/123/skill-creator",
-            )),
-            scope: SkillScope::Builtin,
-            tags: vec![],
-            capabilities: None,
-            compatibility: Default::default(),
-            source: SkillContentSource::File(std::path::PathBuf::from(
-                "/private/tmp/alan/builtin-skill-packages/0.1.0/123/skill-creator/SKILL.md",
-            )),
-            enabled: true,
-            allow_implicit_invocation: true,
-            alan_metadata: Default::default(),
-            compatible_metadata: Default::default(),
-            execution: ResolvedSkillExecution::Delegate {
-                target: "skill-creator".to_string(),
-                source: SkillExecutionResolutionSource::ExplicitMetadata,
-            },
-        }];
-
-        let list = render_skills_list(&skills, true).unwrap();
-        assert!(list.contains("  skill_source: builtin_capability_package"));
-        assert!(list.contains("  execution: delegate(target=skill-creator)"));
-        assert!(list.contains("invoke_delegated_skill"));
-        assert!(!list.contains("skill_path:"));
-        assert!(!list.contains("builtin-skill-packages"));
-    }
-
-    #[test]
-    fn test_render_skills_list_builtin_delegated_skill_degrades_without_path_leak() {
-        let skills = vec![SkillMetadata {
-            id: "skill-creator".to_string(),
-            package_id: Some("builtin:alan-skill-creator".to_string()),
-            name: "Skill Creator".to_string(),
-            description: "Create or update alan skill packages".to_string(),
-            short_description: None,
-            path: std::path::PathBuf::from(
-                "/private/tmp/alan/builtin-skill-packages/0.1.0/123/skill-creator/SKILL.md",
-            ),
-            package_root: Some(std::path::PathBuf::from(
-                "/private/tmp/alan/builtin-skill-packages/0.1.0/123/skill-creator",
-            )),
-            resource_root: Some(std::path::PathBuf::from(
-                "/private/tmp/alan/builtin-skill-packages/0.1.0/123/skill-creator",
-            )),
-            scope: SkillScope::Builtin,
-            tags: vec![],
-            capabilities: None,
-            compatibility: Default::default(),
-            source: SkillContentSource::File(std::path::PathBuf::from(
-                "/private/tmp/alan/builtin-skill-packages/0.1.0/123/skill-creator/SKILL.md",
-            )),
-            enabled: true,
-            allow_implicit_invocation: true,
-            alan_metadata: Default::default(),
-            compatible_metadata: Default::default(),
-            execution: ResolvedSkillExecution::Delegate {
-                target: "skill-creator".to_string(),
-                source: SkillExecutionResolutionSource::ExplicitMetadata,
-            },
-        }];
-
-        let list = render_skills_list(&skills, false).unwrap();
-        assert!(list.contains("  skill_source: builtin_capability_package"));
-        assert!(list.contains("this runtime cannot delegate the builtin capability directly"));
-        assert!(!list.contains("invoke_delegated_skill"));
-        assert!(!list.contains("skill_path:"));
-        assert!(!list.contains("builtin-skill-packages"));
-    }
-
-    #[test]
-    fn test_active_skill_context_hides_builtin_package_paths() {
-        let skill = Skill {
-            metadata: SkillMetadata {
-                id: "memory".to_string(),
-                package_id: Some("builtin:alan-memory".to_string()),
-                name: "Memory".to_string(),
-                description: "Persistent memory across Agent Processes".to_string(),
-                short_description: None,
-                path: std::path::PathBuf::from(
-                    "/private/tmp/alan/builtin-skill-packages/0.1.0/123/memory/SKILL.md",
-                ),
-                package_root: Some(std::path::PathBuf::from(
-                    "/private/tmp/alan/builtin-skill-packages/0.1.0/123/memory",
-                )),
-                resource_root: Some(std::path::PathBuf::from(
-                    "/private/tmp/alan/builtin-skill-packages/0.1.0/123/memory",
-                )),
-                scope: SkillScope::Builtin,
-                tags: vec![],
-                capabilities: None,
-                compatibility: Default::default(),
-                source: SkillContentSource::File(std::path::PathBuf::from(
-                    "/private/tmp/alan/builtin-skill-packages/0.1.0/123/memory/SKILL.md",
-                )),
-                enabled: true,
-                allow_implicit_invocation: true,
-                alan_metadata: Default::default(),
-                compatible_metadata: Default::default(),
-                execution: Default::default(),
-            },
-            content: "# Instructions\nPersist durable context.".to_string(),
-            frontmatter: SkillFrontmatter {
-                name: "Memory".to_string(),
-                description: "Persistent memory across Agent Processes".to_string(),
-                metadata: Default::default(),
-                capabilities: Default::default(),
-                compatibility: Default::default(),
-            },
-        };
-        let envelope = ActiveSkillEnvelope::available(
-            skill.metadata.clone(),
-            SkillActivationReason::ExplicitMention {
-                mention: "memory".to_string(),
-            },
-        );
-
-        let rendered = render_active_skill_prompt_for_runtime(&skill, &envelope, false).rendered;
-        assert!(rendered.contains("canonical_path: builtin:memory"));
-        assert!(rendered.contains("package_root: <builtin capability package>"));
-        assert!(rendered.contains("resource_root: <builtin capability package>"));
-        assert!(rendered.contains("Do not use tools to open builtin package files by path."));
-        assert!(!rendered.contains("builtin-skill-packages"));
-    }
-
-    #[test]
-    fn test_extract_mentions_edge_cases() {
-        // Empty input
-        assert!(extract_mentions("").is_empty());
-
-        // Only $ sign
-        assert!(extract_mentions("$").is_empty());
-
-        // $ at end
-        assert!(extract_mentions("text $").is_empty());
-
-        // Duplicate mentions (should dedupe)
-        assert_eq!(extract_mentions("$skill-a and $skill-a"), vec!["skill-a"]);
-
-        // Legacy underscore separator is rejected
-        assert!(extract_mentions("$skill_name").is_empty());
-
-        // Legacy dot separator is rejected
-        assert!(extract_mentions("$repo.review").is_empty());
-    }
-
-    #[test]
-    fn test_extract_mentions_multiple_same_and_different() {
-        // Multiple skills with duplicates in various positions
-        let mentions = extract_mentions("$skill-a $skill-b $skill-a $skill-c $skill-b");
-        assert_eq!(mentions, vec!["skill-a", "skill-b", "skill-c"]);
-    }
-
-    #[test]
-    fn test_extract_mentions_with_numbers() {
-        assert_eq!(extract_mentions("Use $skill-123"), vec!["skill-123"]);
-        assert!(extract_mentions("$test-v2.0").is_empty());
-        assert_eq!(extract_mentions("Use $skill-name."), vec!["skill-name"]);
-    }
-
-    #[test]
-    fn test_inject_skills_empty() {
-        let result = inject_skills(&[]);
-        assert!(result.is_empty());
-    }
-
-    #[test]
-    fn test_inject_skills_with_resources() {
-        let temp = tempfile::tempdir().unwrap();
-        let skill_dir = temp.path().join("test-skill");
-        std::fs::create_dir(&skill_dir).unwrap();
-        std::fs::create_dir(skill_dir.join("scripts")).unwrap();
-        std::fs::create_dir(skill_dir.join("references")).unwrap();
-        std::fs::create_dir(skill_dir.join("assets")).unwrap();
-
-        // Create resource files
-        std::fs::write(skill_dir.join("scripts/test.sh"), "#!/bin/bash").unwrap();
-        std::fs::write(skill_dir.join("references/ref.md"), "# Reference").unwrap();
-        std::fs::write(skill_dir.join("assets/logo.png"), [0_u8, 159, 146, 150]).unwrap();
-
-        let skill = Skill {
-            metadata: SkillMetadata {
-                id: "test-res".to_string(),
-                package_id: None,
-                name: "Test Resource Skill".to_string(),
-                description: "A test".to_string(),
-                short_description: None,
-                path: skill_dir.join("SKILL.md"),
-                package_root: Some(skill_dir.clone()),
-                resource_root: Some(skill_dir.clone()),
-                scope: SkillScope::Installed,
-                tags: vec![],
-                capabilities: Some(SkillCapabilities {
-                    disclosure: DisclosureConfig {
-                        level3: Level3Resources {
-                            assets: vec!["logo.png".to_string()],
-                            ..Default::default()
-                        },
-                        ..Default::default()
-                    },
-                    ..Default::default()
-                }),
-                compatibility: Default::default(),
-                source: SkillContentSource::File(skill_dir.join("SKILL.md")),
-                enabled: true,
-                allow_implicit_invocation: true,
-                alan_metadata: Default::default(),
-                compatible_metadata: Default::default(),
-                execution: Default::default(),
-            },
-            content: "Read `references/ref.md` before running `scripts/test.sh`.".to_string(),
-            frontmatter: SkillFrontmatter {
-                name: "Test Resource Skill".to_string(),
-                description: "A test".to_string(),
-                metadata: Default::default(),
-                capabilities: Default::default(),
-                compatibility: Default::default(),
-            },
-        };
-
-        let injected = inject_skills(&[skill]);
-        assert!(injected.contains("## Skill: Test Resource Skill"));
-        assert!(injected.contains("### alan Runtime Context"));
-        assert!(injected.contains("### Disclosed Resources"));
-        assert!(injected.contains("#### script: scripts/test.sh"));
-        assert!(injected.contains("#!/bin/bash"));
-        assert!(injected.contains("#### reference: references/ref.md"));
-        assert!(injected.contains("# Reference"));
-        assert!(!injected.contains("#### asset: assets/logo.png"));
-    }
-
-    #[test]
-    fn test_inject_skills_only_expands_declared_resources_when_level2_references_them() {
-        let temp = tempfile::tempdir().unwrap();
-        let skill_dir = temp.path().join("test-skill");
-        std::fs::create_dir(&skill_dir).unwrap();
-        std::fs::create_dir(skill_dir.join("references")).unwrap();
-
-        std::fs::write(skill_dir.join("references/quickstart.md"), "# Quickstart").unwrap();
-
-        let skill = Skill {
-            metadata: SkillMetadata {
-                id: "test-res".to_string(),
-                package_id: None,
-                name: "Test Resource Skill".to_string(),
-                description: "A test".to_string(),
-                short_description: None,
-                path: skill_dir.join("SKILL.md"),
-                package_root: Some(skill_dir.clone()),
-                resource_root: Some(skill_dir.clone()),
-                scope: SkillScope::Installed,
-                tags: vec![],
-                capabilities: Some(SkillCapabilities {
-                    disclosure: DisclosureConfig {
-                        level3: Level3Resources {
-                            references: vec!["quickstart.md".to_string()],
-                            ..Default::default()
-                        },
-                        ..Default::default()
-                    },
-                    ..Default::default()
-                }),
-                compatibility: Default::default(),
-                source: SkillContentSource::File(skill_dir.join("SKILL.md")),
-                enabled: true,
-                allow_implicit_invocation: true,
-                alan_metadata: Default::default(),
-                compatible_metadata: Default::default(),
-                execution: Default::default(),
-            },
-            content: "Read `quickstart.md` before using this skill.".to_string(),
-            frontmatter: SkillFrontmatter {
-                name: "Test Resource Skill".to_string(),
-                description: "A test".to_string(),
-                metadata: Default::default(),
-                capabilities: Default::default(),
-                compatibility: Default::default(),
-            },
-        };
-
-        let injected = inject_skills(&[skill]);
-        assert!(injected.contains("#### reference: references/quickstart.md"));
-        assert!(injected.contains("# Quickstart"));
-    }
-
-    #[test]
-    fn test_inject_skills_matches_declared_resources_with_fragment_or_query_suffixes() {
-        let temp = tempfile::tempdir().unwrap();
-        let skill_dir = temp.path().join("test-skill");
-        std::fs::create_dir(&skill_dir).unwrap();
-        std::fs::create_dir(skill_dir.join("references")).unwrap();
-
-        std::fs::write(skill_dir.join("references/quickstart.md"), "# Quickstart").unwrap();
-
-        let build_skill = |content: &str| Skill {
-            metadata: SkillMetadata {
-                id: "test-res".to_string(),
-                package_id: None,
-                name: "Test Resource Skill".to_string(),
-                description: "A test".to_string(),
-                short_description: None,
-                path: skill_dir.join("SKILL.md"),
-                package_root: Some(skill_dir.clone()),
-                resource_root: Some(skill_dir.clone()),
-                scope: SkillScope::Installed,
-                tags: vec![],
-                capabilities: Some(SkillCapabilities {
-                    disclosure: DisclosureConfig {
-                        level3: Level3Resources {
-                            references: vec!["quickstart.md".to_string()],
-                            ..Default::default()
-                        },
-                        ..Default::default()
-                    },
-                    ..Default::default()
-                }),
-                compatibility: Default::default(),
-                source: SkillContentSource::File(skill_dir.join("SKILL.md")),
-                enabled: true,
-                allow_implicit_invocation: true,
-                alan_metadata: Default::default(),
-                compatible_metadata: Default::default(),
-                execution: Default::default(),
-            },
-            content: content.to_string(),
-            frontmatter: SkillFrontmatter {
-                name: "Test Resource Skill".to_string(),
-                description: "A test".to_string(),
-                metadata: Default::default(),
-                capabilities: Default::default(),
-                compatibility: Default::default(),
-            },
-        };
-
-        let fragment_injected = inject_skills(&[build_skill(
-            "Read `quickstart.md#setup` before using this skill.",
-        )]);
-        assert!(fragment_injected.contains("#### reference: references/quickstart.md"));
-
-        let query_injected = inject_skills(&[build_skill(
-            "Read `quickstart.md?view=plain` before using this skill.",
-        )]);
-        assert!(query_injected.contains("#### reference: references/quickstart.md"));
-    }
-
-    #[test]
-    fn test_inject_skills_matches_prefixed_declared_resources_from_bare_references() {
-        let temp = tempfile::tempdir().unwrap();
-        let skill_dir = temp.path().join("test-skill");
-        std::fs::create_dir(&skill_dir).unwrap();
-        std::fs::create_dir(skill_dir.join("references")).unwrap();
-
-        std::fs::write(skill_dir.join("references/quickstart.md"), "# Quickstart").unwrap();
-
-        let skill = Skill {
-            metadata: SkillMetadata {
-                id: "test-res".to_string(),
-                package_id: None,
-                name: "Test Resource Skill".to_string(),
-                description: "A test".to_string(),
-                short_description: None,
-                path: skill_dir.join("SKILL.md"),
-                package_root: Some(skill_dir.clone()),
-                resource_root: Some(skill_dir.clone()),
-                scope: SkillScope::Installed,
-                tags: vec![],
-                capabilities: Some(SkillCapabilities {
-                    disclosure: DisclosureConfig {
-                        level3: Level3Resources {
-                            references: vec!["references/quickstart.md".to_string()],
-                            ..Default::default()
-                        },
-                        ..Default::default()
-                    },
-                    ..Default::default()
-                }),
-                compatibility: Default::default(),
-                source: SkillContentSource::File(skill_dir.join("SKILL.md")),
-                enabled: true,
-                allow_implicit_invocation: true,
-                alan_metadata: Default::default(),
-                compatible_metadata: Default::default(),
-                execution: Default::default(),
-            },
-            content: "Read `quickstart.md` before using this skill.".to_string(),
-            frontmatter: SkillFrontmatter {
-                name: "Test Resource Skill".to_string(),
-                description: "A test".to_string(),
-                metadata: Default::default(),
-                capabilities: Default::default(),
-                compatibility: Default::default(),
-            },
-        };
-
-        let injected = inject_skills(&[skill]);
-        assert!(injected.contains("#### reference: references/quickstart.md"));
-        assert!(injected.contains("# Quickstart"));
-    }
-
-    #[test]
-    fn test_declared_resource_reference_candidates_normalize_windows_separators() {
-        let candidates = declared_resource_reference_candidates(
-            SkillResourceKind::Reference,
-            "guides/setup.md",
-            r"references\guides\setup.md",
-        );
-
-        assert!(candidates.contains(&"references/guides/setup.md".to_string()));
-        assert!(content_contains_resource_reference(
-            "Read `references/guides/setup.md` before running the skill.",
-            "references/guides/setup.md",
-        ));
-    }
-
-    #[test]
-    fn test_inject_skills_uses_custom_level2_file() {
-        let temp = tempfile::tempdir().unwrap();
-        let skill_dir = temp.path().join("test-skill");
-        std::fs::create_dir(&skill_dir).unwrap();
-        std::fs::write(skill_dir.join("details.md"), "Expanded instructions.").unwrap();
-
-        let skill = Skill {
-            metadata: SkillMetadata {
-                id: "test-res".to_string(),
-                package_id: None,
-                name: "Test Resource Skill".to_string(),
-                description: "A test".to_string(),
-                short_description: None,
-                path: skill_dir.join("SKILL.md"),
-                package_root: Some(skill_dir.clone()),
-                resource_root: Some(skill_dir.clone()),
-                scope: SkillScope::Installed,
-                tags: vec![],
-                capabilities: Some(SkillCapabilities {
-                    disclosure: DisclosureConfig {
-                        level2: "details.md".to_string(),
-                        ..Default::default()
-                    },
-                    ..Default::default()
-                }),
-                compatibility: Default::default(),
-                source: SkillContentSource::File(skill_dir.join("SKILL.md")),
-                enabled: true,
-                allow_implicit_invocation: true,
-                alan_metadata: Default::default(),
-                compatible_metadata: Default::default(),
-                execution: Default::default(),
-            },
-            content: "Fallback instructions.".to_string(),
-            frontmatter: SkillFrontmatter {
-                name: "Test Resource Skill".to_string(),
-                description: "A test".to_string(),
-                metadata: Default::default(),
-                capabilities: Default::default(),
-                compatibility: Default::default(),
-            },
-        };
-
-        let injected = inject_skills(&[skill]);
-        assert!(injected.contains("source: details.md"));
-        assert!(injected.contains("Expanded instructions."));
-        assert!(!injected.contains("Fallback instructions."));
-    }
-
-    #[test]
-    fn test_descriptor_skill_disclosure_reads_only_its_file_tree() {
-        let namespace_root = std::path::PathBuf::from("/lib/pkg/example/skills/review");
-        let tree = crate::ProcessFileTree::new(std::collections::BTreeMap::from([
-            ("SKILL.md".to_string(), b"Fallback instructions.".to_vec()),
-            (
-                "details.md".to_string(),
-                b"Read `references/guide.md` before reviewing.".to_vec(),
-            ),
-            (
-                "references/guide.md".to_string(),
-                b"# Descriptor Guide".to_vec(),
-            ),
-        ]))
-        .unwrap();
-        let skill = Skill {
-            metadata: SkillMetadata {
-                id: "review".to_string(),
-                package_id: Some("example".to_string()),
-                name: "Review".to_string(),
-                description: "Review a change".to_string(),
-                short_description: None,
-                path: namespace_root.join("SKILL.md"),
-                package_root: Some(namespace_root.clone()),
-                resource_root: Some(namespace_root),
-                scope: SkillScope::Descriptor,
-                tags: vec![],
-                capabilities: Some(SkillCapabilities {
-                    disclosure: DisclosureConfig {
-                        level2: "details.md".to_string(),
-                        ..Default::default()
-                    },
-                    ..Default::default()
-                }),
-                compatibility: Default::default(),
-                source: SkillContentSource::Descriptor {
-                    content: std::sync::Arc::from("Fallback instructions."),
-                    file_tree: tree,
-                },
-                enabled: true,
-                allow_implicit_invocation: true,
-                alan_metadata: Default::default(),
-                compatible_metadata: Default::default(),
-                execution: Default::default(),
-            },
-            content: "Fallback instructions.".to_string(),
-            frontmatter: SkillFrontmatter {
-                name: "Review".to_string(),
-                description: "Review a change".to_string(),
-                metadata: Default::default(),
-                capabilities: Default::default(),
-                compatibility: Default::default(),
-            },
-        };
-
-        let injected = inject_skills(&[skill]);
-
-        assert!(injected.contains("source: details.md"));
-        assert!(injected.contains("Read `references/guide.md` before reviewing."));
-        assert!(injected.contains("#### reference: references/guide.md"));
-        assert!(injected.contains("# Descriptor Guide"));
-        assert!(!injected.contains("Fallback instructions."));
-    }
-
-    #[test]
-    fn test_load_disclosed_text_content_caps_large_files_by_byte_budget() {
-        let temp = tempfile::tempdir().unwrap();
-        let path = temp.path().join("large.txt");
-        std::fs::write(
-            &path,
-            "a".repeat(MAX_DISCLOSED_RESOURCE_BYTES as usize + 1024),
-        )
-        .unwrap();
-
-        let content = load_disclosed_text_content(
-            &path,
-            MAX_DISCLOSED_RESOURCE_BYTES,
-            Some(MAX_DISCLOSED_RESOURCE_CHARS),
-        )
-        .unwrap();
-
-        assert!(content.starts_with('a'));
-        assert!(content.contains(&format!("exceeded {MAX_DISCLOSED_RESOURCE_BYTES} bytes")));
-    }
-
-    #[test]
-    fn test_extract_resource_references_ignores_urls_and_trims_punctuation() {
-        let references = extract_resource_references(
-            "Use https://example.com/scripts/setup.sh, then read references/guide.md.",
-        );
-
-        assert_eq!(references, vec!["references/guide.md"]);
-    }
-
-    #[test]
-    fn test_inject_skills_extracts_sentence_refs_and_dot_slash_paths() {
-        let temp = tempfile::tempdir().unwrap();
-        let skill_dir = temp.path().join("test-skill");
-        std::fs::create_dir(&skill_dir).unwrap();
-        std::fs::create_dir(skill_dir.join("scripts")).unwrap();
-        std::fs::create_dir(skill_dir.join("references")).unwrap();
-        std::fs::write(
-            skill_dir.join("scripts/setup.sh"),
-            "#!/bin/bash\necho setup",
-        )
-        .unwrap();
-        std::fs::write(skill_dir.join("references/guide.md"), "# Guide").unwrap();
-
-        let skill = Skill {
-            metadata: SkillMetadata {
-                id: "test-res".to_string(),
-                package_id: None,
-                name: "Test Resource Skill".to_string(),
-                description: "A test".to_string(),
-                short_description: None,
-                path: skill_dir.join("SKILL.md"),
-                package_root: Some(skill_dir.clone()),
-                resource_root: Some(skill_dir.clone()),
-                scope: SkillScope::Installed,
-                tags: vec![],
-                capabilities: None,
-                compatibility: Default::default(),
-                source: SkillContentSource::File(skill_dir.join("SKILL.md")),
-                enabled: true,
-                allow_implicit_invocation: true,
-                alan_metadata: Default::default(),
-                compatible_metadata: Default::default(),
-                execution: Default::default(),
-            },
-            content: "Use https://example.com/scripts/setup.sh, then read references/guide.md. After that run ./scripts/setup.sh."
-                .to_string(),
-            frontmatter: SkillFrontmatter {
-                name: "Test Resource Skill".to_string(),
-                description: "A test".to_string(),
-                metadata: Default::default(),
-                capabilities: Default::default(),
-                compatibility: Default::default(),
-            },
-        };
-
-        let injected = inject_skills(&[skill]);
-        assert!(injected.contains("#### reference: references/guide.md"));
-        assert!(injected.contains("#### script: scripts/setup.sh"));
-        assert_eq!(injected.matches("#### script: scripts/setup.sh").count(), 1);
-    }
-
-    #[test]
-    fn test_materialize_disclosed_resources_loads_only_capped_selection() {
-        let load_count = std::cell::Cell::new(0);
-        let resources = (0..12).map(|index| PendingDisclosedSkillResource {
-            kind: SkillResourceKind::Reference,
-            display_path: format!("references/ref-{index:02}.md"),
-            path: PathBuf::from(format!("/tmp/ref-{index:02}.md")),
-        });
-
-        let loaded = materialize_disclosed_resources(resources, |_| {
-            load_count.set(load_count.get() + 1);
-            Some("content".to_string())
-        });
-
-        assert_eq!(loaded.len(), MAX_DISCLOSED_RESOURCE_COUNT);
-        assert_eq!(load_count.get(), MAX_DISCLOSED_RESOURCE_COUNT);
-        assert_eq!(loaded[0].display_path, "references/ref-00.md");
-        assert_eq!(loaded[7].display_path, "references/ref-07.md");
-    }
-
-    #[test]
-    fn test_inject_skills_caps_large_level2_file_by_byte_budget() {
-        let temp = tempfile::tempdir().unwrap();
-        let skill_dir = temp.path().join("test-skill");
-        std::fs::create_dir(&skill_dir).unwrap();
-        std::fs::write(
-            skill_dir.join("details.md"),
-            "b".repeat(MAX_DISCLOSED_LEVEL2_BYTES as usize + 1024),
-        )
-        .unwrap();
-
-        let skill = Skill {
-            metadata: SkillMetadata {
-                id: "test-res".to_string(),
-                package_id: None,
-                name: "Test Resource Skill".to_string(),
-                description: "A test".to_string(),
-                short_description: None,
-                path: skill_dir.join("SKILL.md"),
-                package_root: Some(skill_dir.clone()),
-                resource_root: Some(skill_dir.clone()),
-                scope: SkillScope::Installed,
-                tags: vec![],
-                capabilities: Some(SkillCapabilities {
-                    disclosure: DisclosureConfig {
-                        level2: "details.md".to_string(),
-                        ..Default::default()
-                    },
-                    ..Default::default()
-                }),
-                compatibility: Default::default(),
-                source: SkillContentSource::File(skill_dir.join("SKILL.md")),
-                enabled: true,
-                allow_implicit_invocation: true,
-                alan_metadata: Default::default(),
-                compatible_metadata: Default::default(),
-                execution: Default::default(),
-            },
-            content: "Fallback instructions.".to_string(),
-            frontmatter: SkillFrontmatter {
-                name: "Test Resource Skill".to_string(),
-                description: "A test".to_string(),
-                metadata: Default::default(),
-                capabilities: Default::default(),
-                compatibility: Default::default(),
-            },
-        };
-
-        let injected = inject_skills(&[skill]);
-        assert!(injected.contains("source: details.md"));
-        assert!(injected.contains(&format!("exceeded {MAX_DISCLOSED_LEVEL2_BYTES} bytes")));
-        assert!(!injected.contains("Fallback instructions."));
-    }
-
-    #[test]
-    fn test_format_disclosed_resources_uses_safe_fence_for_embedded_backticks() {
-        let rendered = format_disclosed_resources(&[DisclosedSkillResource {
-            kind: SkillResourceKind::Reference,
-            display_path: "references/ref.md".to_string(),
-            tracked_path: PromptTrackedPath::prefix_bytes(PathBuf::from("/tmp/ref.md"), 16),
-            content: Some("before\n```md\ninside\n```\nafter".to_string()),
-        }]);
-
-        assert!(rendered.contains("````md"));
-        assert!(rendered.contains("\n````"));
-    }
-
-    #[test]
-    fn test_inject_skills_no_parent_path() {
-        // Test the edge case where skill path has no parent
-        let skill = Skill {
-            metadata: SkillMetadata {
-                id: "no-parent".to_string(),
-                package_id: None,
-                name: "No Parent".to_string(),
-                description: "Test".to_string(),
-                short_description: None,
-                path: std::path::PathBuf::from("SKILL.md"), // No parent
-                package_root: None,
-                resource_root: None,
-                scope: SkillScope::Installed,
-                tags: vec![],
-                capabilities: None,
-                compatibility: Default::default(),
-                source: SkillContentSource::File(std::path::PathBuf::from("SKILL.md")),
-                enabled: true,
-                allow_implicit_invocation: true,
-                alan_metadata: Default::default(),
-                compatible_metadata: Default::default(),
-                execution: Default::default(),
-            },
-            content: "Content".to_string(),
-            frontmatter: SkillFrontmatter {
-                name: "No Parent".to_string(),
-                description: "Test".to_string(),
-                metadata: Default::default(),
-                capabilities: Default::default(),
-                compatibility: Default::default(),
-            },
-        };
-
-        let injected = inject_skills(&[skill]);
-        assert!(injected.contains("## Skill: No Parent"));
-        // Should not panic and should not have Resources section
-        assert!(!injected.contains("### Disclosed Resources"));
-    }
-
-    #[test]
-    fn test_build_prompt_with_skills_empty() {
-        let prompt = build_prompt_with_skills("Just user input", &[]);
-        assert_eq!(prompt, "Just user input");
-    }
-
-    #[test]
-    fn test_render_skills_list_empty() {
-        assert!(render_skills_list(&[], true).is_none());
-    }
-
-    #[test]
-    fn test_render_skill_not_found_with_similar() {
-        let available = vec![
-            SkillMetadata {
-                id: "test-skill".to_string(),
-                package_id: None,
-                name: "Test Skill".to_string(),
-                description: "A test".to_string(),
-                short_description: None,
-                path: std::path::PathBuf::from("/test/SKILL.md"),
-                package_root: None,
-                resource_root: None,
-                scope: SkillScope::Installed,
-                tags: vec![],
-                capabilities: None,
-                compatibility: Default::default(),
-                source: SkillContentSource::File(std::path::PathBuf::from("/test/SKILL.md")),
-                enabled: true,
-                allow_implicit_invocation: true,
-                alan_metadata: Default::default(),
-                compatible_metadata: Default::default(),
-                execution: Default::default(),
-            },
-            SkillMetadata {
-                id: "testing".to_string(),
-                package_id: None,
-                name: "Testing".to_string(),
-                description: "Testing skill".to_string(),
-                short_description: None,
-                path: std::path::PathBuf::from("/testing/SKILL.md"),
-                package_root: None,
-                resource_root: None,
-                scope: SkillScope::Installed,
-                tags: vec![],
-                capabilities: None,
-                compatibility: Default::default(),
-                source: SkillContentSource::File(std::path::PathBuf::from("/testing/SKILL.md")),
-                enabled: true,
-                allow_implicit_invocation: true,
-                alan_metadata: Default::default(),
-                compatible_metadata: Default::default(),
-                execution: Default::default(),
-            },
-        ];
-
-        let msg = render_skill_not_found("test", &available);
-        assert!(msg.contains("Skill '$test' not found"));
-        assert!(msg.contains("Did you mean:"));
-        assert!(msg.contains("$test-skill"));
-    }
-
-    #[test]
-    fn test_render_skill_not_found_no_similar() {
-        let available = vec![SkillMetadata {
-            id: "other".to_string(),
-            package_id: None,
-            name: "Other".to_string(),
-            description: "Other skill".to_string(),
-            short_description: None,
-            path: std::path::PathBuf::from("/other/SKILL.md"),
-            package_root: None,
-            resource_root: None,
-            scope: SkillScope::Installed,
-            tags: vec![],
-            capabilities: None,
-            compatibility: Default::default(),
-            source: SkillContentSource::File(std::path::PathBuf::from("/other/SKILL.md")),
-            enabled: true,
-            allow_implicit_invocation: true,
-            alan_metadata: Default::default(),
-            compatible_metadata: Default::default(),
-            execution: Default::default(),
-        }];
-
-        let msg = render_skill_not_found("xyz", &available);
-        assert!(msg.contains("Skill '$xyz' not found"));
-        assert!(msg.contains("Use `/skills` to see available skills"));
-        assert!(!msg.contains("Did you mean:"));
-    }
-
-    #[test]
-    fn test_render_skill_not_found_partial_match() {
-        // Test when the mention contains the skill id
-        let available = vec![SkillMetadata {
-            id: "rust".to_string(),
-            package_id: None,
-            name: "Rust".to_string(),
-            description: "Rust skill".to_string(),
-            short_description: None,
-            path: std::path::PathBuf::from("/rust/SKILL.md"),
-            package_root: None,
-            resource_root: None,
-            scope: SkillScope::Installed,
-            tags: vec![],
-            capabilities: None,
-            compatibility: Default::default(),
-            source: SkillContentSource::File(std::path::PathBuf::from("/rust/SKILL.md")),
-            enabled: true,
-            allow_implicit_invocation: true,
-            alan_metadata: Default::default(),
-            compatible_metadata: Default::default(),
-            execution: Default::default(),
-        }];
-
-        // "rustacean" contains "rust"
-        let msg = render_skill_not_found("rustacean", &available);
-        assert!(msg.contains("Did you mean:"));
-        assert!(msg.contains("$rust"));
-    }
-}
+#[path = "injector_prompt_tests.rs"]
+mod prompt_tests;
+
+#[cfg(test)]
+#[path = "injector_resource_tests.rs"]
+mod resource_tests;
+
+#[cfg(test)]
+#[path = "injector_message_tests.rs"]
+mod message_tests;
